Added new menu items to the menubar.
Refactored the menu's ETS table to introduce menu groups, which allow groups of menu items to be enabled/disabled as a single unit. A menu item is associated with 0 or more groups. A group is defined in ide.hrl, as a bit flag. The toggle_menu_group() function can then be called providing the bitmask of those groups that are to be affected.