Killed the toggle panes deadlock and show/hide bugs.
The deadlock bug was a result of these events being managed in the ETS table. The function called from the ETS table ide:toggle_pane required much of the state data from ide, which meant it made a wx_object:call to itself, hence deadlock.
The toggle menu events no longer get called from the ETS (apply/3) but are handled using handle_event/2 in ide, and therefore have access to the state required.

I've removed the update_label option from the ETS table in ide_menu, as this is no longer required, because I've replaced the very restrictive update_label/2 function (only dealt specifically with the toggle menu items label) with a more generic function update_label/3 which will update the label of any menu item passed to it. Also added another generic function to toggle enabled on menu items.

The show/hide bug is now fixed too. All show/hide/maximise buttons now produce the correct result. Changing the labels is still an issue as there is no checks as yet for when the user manually drags a sash to the edge, which will result in the wrong label being displayed. We either add these checks or don't change the labels at all and have, for example, "Toggle Utilities" rather than seperate "Show Utilities" and "Hide Utilities". Unitl this is decided, I haven't bothered updating the labels when panes are maximised/minimised.