Added error catching for the port, when it cannot be opened. A placeholder is now inserted where the console should have been with an error message for the user. This allows the ide to stop and start normally on MSW until we get the port working properly cross-platform.
As it stands the port is linked to the ide, and if the port cannot open an throw is thrown which IS NOT yet caught in ide, but rather the EXIT signal trapped.
The calling process (ide) knows about the error because the link() function fails with a badarg in port:start/0.