Refactored console_port, making the module fully functional on MSW. Renamed port to console_port in anticipation of other ports. console_port now uses a gen_server to manage requests and responses. The port can buffer messages should the receiver not be ready (i.e. during initialisation) and can later be cleared when the receiver becomes ready. This is done using the buffer_responses/1 and flush_buffer functions.
The current implementation of the buffer is a list, and there isn't yet any
way to discard unwanted messages, as they are sent directly to console_wx (was ide_shell) for displaying to the user via the console_parser (was parser).

There isn't yet a supervisor to restart the console port should it exit.

Added properly formatted commenting to all console files.