Started refactoring all project related funtions into their own module, removing them from doc_manager where they currently belong.
Each project is now assigned a unique id when created/opened, and any data associated to that project is accessible through its id.

The project_tree will no longer be the module that stores data related open projects, and can concentrate instead on just building and displaying the treectrl. The new module project_manager is the point of access to any data relating to projects, including any open project documents.