Single line indenting now affects the whole line, and the position of the caret is irrelevant.
The auto-indent function has been improved. It now happens within a sync event so that the caret can be moved prior to other events firing. This stops the ugly flickering of the caret. The indent will be triggered if the previous line contains a keyword such as case/receive/fun etc.
Currently no un-indenting is implemented.

Using sync events has stopped the bug whereby if the user was to hold down RETURN in the editor, the rate at which lines were being created became out of sync with the position of the character, so for example after reaching line 100 (line number margin), the caret might be on line 95. This was also illustrated by having the RETURN key print a character to the editor and viewing the positoin that the character appeared whilst RETURN was being pressed. Some lines would have no character printed while others might have two or more.

The comment function when applied to a single line will now not ignore any %'s inserted by the user. Rather than look for a comment character at position 0 in the line (where the program inserts them), it now looks for the first non-whitespace character.