Directory structure changes.
Changed the close_document implementation, which now makes a cast after vetoing the close tab event (from the close button on the tab).

The previous implementation performed all of the necessary functions within a callback. The callbacks are always invoked by another process and have exclusive usage of the gui when invoked. This means that a callback fun can not use the process dictionary and should not make calls to other processes. Calls to another process inside a callback fun may cause a deadlock if the other process is waiting on completion of his call to the gui.

To solve this problem, which was causing deadlock, the callback simply makes a cast to the doc_manager and then terminates, ao all processing is then done inside the doc_manager process.

We haven't the implementation for close_all i.e. should it display the save changes dialog for each modified document, or should it close only unmodified documents. The implementation for saving files that have no associated path is also undefined. It is still possible to create a untitled document by double clicking the notebook tabbed area, which bypasses the new file dialog. Whether we choose to allow this or not will influence how/whether we need an implemention for the save_new_document which avoids the deadlock problems described above, from which it currently suffers (save_new_document makes a wx_object:call from within the close event).

Added but yet to test the supervisor for the console_port process.