Refactored log.erl to remove duplicate code and generally tidy it up.
The compiler_port.erl module has been refactored. Both start functions now expect Cwd (current working directory) as the first parameter, with start/2 accepting an optional list of flags most of which are sent to erlc. If the 'file' flag is set in Config then Cwd is the absolute path of the file to compile, otherwise it
is the root directory that contains all erlang source files to be compiled.

Currently the possible flags are supress_warnings, verbose, smp, makefile, all_errors which are mostly self explanatory. 'makefile' creates a makefile, and all_errors converts all warnings to errors.