Console refactoring.
console_port

Buffering/flushing responses no longer required, as we have added the ability to suppress the response when
calling the function eval/1|eval/2 was call_port/1. No longer then, do we have to buffer
messages whilst the actual console text ctrl (console) is initalised, we simply discard them.
eval/2 accepts a boolean flag to switch off the response. The default behaviour is true (keep
response). The flag is not currently synchronized and therefore should a user make a call to the port
which takes a while to complete, and in the meantime makes other calls in which they switch off the response,
the response for the first call may also be discarded when it finishes.

console_parser

The parser now makes use of a receive loop to assemble responses into complete units.
Previously each line (at newline) was passed back to the console when received so:
"** exception error: an error occurred when evaluating an arithmetic expression\n
     in operator  +/2\n
                  called as tom + 5\n"
                  was passed back as three lines.
                  The parser now assembles the above into a response (which it is), and passes it back to
                  the console only when complete, i.e. the parser receives a prompt ("4> \n").

                  This refactoring of the console modules allows us to manage the prompt ourselves which
                  gives us much greater flexibility when writing to the console, and allows us to hide
                  certain operations and/or output from the user (i.e cd(), l()). We can also write at any position without interfering with
                  the prompt (command number), which A) looked ugly, and B) became out of sync (jump from 2 to 4, for example, if we deleted a line).
                  The prompt returned from the port is now discarded, as is the newline character at the end of each response (but not those within a long response).

console_wx

We have stopped using the numbered prompt for now, and are using a static prompt defined in a macro.