Removed menu_highlight and menu_close handlers as not used.
Added global handler for the paste function.

Note:

Currently on MSW and Linux the menu event is not caught by the in focus control even when
it has a registered handler for it. On OSX the event is caught as expected (and will
propagate up through the hierarchy). Not sure which of
these behaviours is correct. Due to this we have to catch the copy/paste/cut etc. menu
events here, find out which control was in focus and then execute the correct function on it.
We have tried but failed to find an wxErlang alternative to the wxWidget
wxEventHandler->proccess_event() function that would allow us to pass the event to the in-focus
control, and so this is the workaround we have adopted.