Application resources (such as icons) are now found using their absolute path, located at runtime. Previously if you were in the application directory "ide" and did:
    erl -pa ebin

none of the image resources would be found as erlang was looking for them relative to the CWD. The CWD working directory therefore had to be "ide/ebin". This is no longer the case, and starting the vm with the above command works fine.

The function that calculates the path to the resouce file rc_dir/1 is in ide_lib_widgets.

Renamed the icon folder to rc.

Updated those include statements that used links relative to CWD.