Undo/redo is now only enabled only when an editor instance is in focus, and so targets the editor directly (ticket #13).
Ticket #81. We use the following id's (?wxID_CUT, ?wxID_COPY, ?wxID_PASTE, ?wxID_SELECTALL, ?wxID_DELETE) in the menu item, which attach default handlers to wxTextCtrl only. As a result we have to implement these common functions manually for all wxStyledTextCtrl instances, which I've done. Each wxStyledTextCtrl enables and disables only those Id's which the ctrl implements on set_focus/kill_focus.

Ticket #88. The "Save" menu item is now disabled when a savepoint is reached.

The above covers the majority of ticket #10. Which I will mark as complete, and open a new ticket for enabling/disabling compile etc. based on filetype, and also the saved state of a project.