On entering an invalid path to erl in the preferences the console will be replaced with an error message. The console will be restarted when a valid path to erl is entered. The console supervisor exit is trapped in ide, and the placeholder is inserted. It is restarted with a call to ide:restart_console from the preferences.
Added a new menu group MENU_GROUP_ERL for all menu items that require a working console. Updated the toggle_menu_group function so that any items within the MENU_GROUP_ERL that are toggled by (belong to) other groups, will not be toggled when the console is unavailable. i.e. Opening a new document would normally toggle on the compile file menu item, which is now prevented (when console is unavailable).