Themes
Now uses Notepad++ themes. These don't (usually) have an Erlang LexerType (XML) element, so they need to be transformed to add Erlang support.
For each Erlang style (there are currently 24), the original theme is queried for the nearest match (by name - i.e "COMMENT" or "VARIABLE"), and these are added to a new Erlang LexerType element (node).
If no matching style name exists (i.e. "ATOM"), a suitable alternative can be defined in the theme module. If no alternative is offered then "DEFAULT" is used.
"Import" in the "Theme" menu performs this transformation for the user and writes the theme to the theme directory.

The above means Deverl now supports multiple languages for highlighting (currently Erlang, XML and Default - more can be added easily by defining a new lang_info record in the editor module). The filetype is established when a file is opened, and the correct lexer is then extracted from the theme. Should the filetype not be recognised then the "Default" is used (i.e. no syntax highlighting). The language can be changed on a per document basis from the "Language" menu.

Preferences

Adding a new preference to the source (when a preference file already exists) will no longer cause Deverl to crash on boot (when this new pref is requested). It is retrieved from the pref defaults if the ETS doesn't contain it.

Other minor enhancements/improvements.